+++ Tests/RapidSpecs/ConformancesSpecs/EmptyCollectionConformancesSpec.swift
@@ -20,14 +20,14 @@ import Nimble
 
 final class EmptyCollectionConformancesSpec: QuickSpec {
     override class func spec() {
-        describe("EmptyCollection's conformance to ExpressibleByArrayLiteral") {
+        describe("EmptyCollection's conformance to ExpressibleByEmptyArrayLiteral") {
             it("only allows an empty literal") {
                 // If this compiles, we're good to go.
                 let _: EmptyCollection<Int> = [] // adding an element should throw a compiler error
             }
         }
         
-        describe("EmptyCollection's conformance to ExpressibleByDictionaryLiteral") {
+        describe("EmptyCollection's conformance to ExpressibleByEmptyDictionaryLiteral") {
             it("only allows an empty literal") {
                 // If this compiles, we're good to go.
                 let _: EmptyCollection<Int> = [:] // adding an element should throw a compiler error